+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
-## [2.0.1](https://github.com/contao-themes-net/odd-theme-bundle/tree/2.0.1) – 2022-02-18
+## [2.0.1](https://github.com/contao-themes-net/odd-theme-bundle/tree/2.0.1) – 2022-03-09
 
+- [fix] add support for Symfony 5 public entry point
+- [fix] change bundle namespace
 - [fix] update fe_page template
 
 ## [2.0.0](https://github.com/contao-themes-net/odd-theme-bundle/tree/2.0.0) – 2022-02-17